refactor for better structure
